Use 64-bit integer for number of trials

+++ src/gibbon-analysis-move.h
@@ -95,7 +95,7 @@ struct _GibbonAnalysisMove
         gdouble da_noise;
         gboolean da_use_prune;
         gdouble da_p[2][7];
-        guint da_trials;
+        guint64 da_trials;
         gboolean da_take_analysis;
 
         /*< private >*/
